+++ pyproject.toml
@@ -19,9 +19,11 @@ sections = [
 profile = "black"
 
 [tool.mypy]
-python_version = "3.7"
+python_version = "3.8"
 ignore_missing_imports = "True"
 scripts_are_modules = "True"
+# strict = "True"
+# follow_imports = "skip"
 
 [tool.pyright]
 include = ["src"]
